# Optimized loop bounds to allow the compiler to unroll the loop.