@terrelln 2019-11-05
Awesome! I'll look into benchmarking tomorrow, unless someone merges this before then.