@terrelln 2019-11-08
Sorry, I meant to benchmark this today, but my benchmark machine is turned off! I'll try to get to this tonight.