@bimbashrestha 2019-11-14
I can't reproduce the ~1-2% improvement mentioned on my desktop. I'm using the commit before this patch as the baseline and I get slower results every time when decompressing the compressed 10 KB silesia.tar file.

`head -c 1K silesia.tar > /tmp/test && ./zstd /tmp/test -o /tmp/test.zst`

Baseline: `./zstd -d -b /tmp/test.zst`
`3#test.zst          :      1024 ->       122 (8.393),  0.00 MB/s , 859.8 MB/s`
`3#test.zst          :      1024 ->       122 (8.393),  0.00 MB/s , 859.6 MB/s `
`3#test.zst          :      1024 ->       122 (8.393),  0.00 MB/s , 858.7 MB/s`

This patch: `./zstd -d -b /tmp/test.zst`
`3#test.zst          :      1024 ->       122 (8.393),  0.00 MB/s , 847.3 MB/s`
`3#test.zst          :      1024 ->       122 (8.393),  0.00 MB/s , 846.8 MB/s `
`3#test.zst          :      1024 ->       122 (8.393),  0.00 MB/s , 847.0 MB/s`