@terrelln 2019-11-14
I've measured a performance improvement with clang-9.0.0, and a performance decrease with gcc-9.2.0. Measured with `zstd -B10K -b silesia.tar`

| Compiler | Previous speed | New speed | Change |
| clang-9.0.0 | 917.7 MB/s | 932.9 MB/s | +1.5% |
| gcc-9.2.0 | 993.5 MB/s | 988.6 MB/s | -0.5% |

There is probably a version of this patch that doesn't regress gcc. Are you interested in continuing to work on it?