@legrosbuffle 2019-11-15
Sorry for the delay, and thank you both for investigating. We're using clang, I admit that I did not try with gcc.

I've updated the patch with your unrolled version.